DNS/requirements.txt
dnslib
dnspython
flask
pyopenssl
